+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 #### New Features
 
 * [#91](https://github.com/SamR1/FitTrackee/issues/91) - Display elevation chart with min and max altitude of workout
+* [#90](https://github.com/SamR1/FitTrackee/issues/90) - Add user sports preferences
 * [#18](https://github.com/SamR1/FitTrackee/issues/18) - Better UI
 
 #### Bugs Fixed
@@ -21,6 +22,8 @@
 
 * [#98/#109](https://github.com/SamR1/FitTrackee/pull/109) - Added stopped_speed_threshold to support slow movement
 * [#84/#93](https://github.com/SamR1/FitTrackee/pull/93) - Add elevation data and new sports
+
+In this release 5 issue were closed.
 
 
 ## Version 0.4.9 (2021/07/16)